Implement software watchdog
Add fields "watchdog_remaining_ticks" and "watchdog_interval_ticks" to
Tcb struct.
Task can subscribe to the watchdog via "watchdog_subscribe" syscall.
With each systick "remaining_ticks" will be decremented for subscribed
READY tasks. If this counter drops to zero, kernel will panic.
Tasks can reset the counter via "watchdog_feed" syscall as well as
unsubscribe via "watchdog_unsubscribe".

+++ whyos/src/task/stack.rs
@@ -67,7 +67,7 @@ impl HwStackFrame {
             r3: 0x33333333,
             r12: 0xCCCCCCCC,
             lr: return_handler as u32, // on exception return, use psp in thread mode
-            pc: entry_point as usize as u32,
+            pc: entry_point as u32,
             xpsr: XPSR_THUMB, // thumb bit, must be set for cortex-m
         }
     }
